requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.33.0
sqlalchemy==2.0.41
pydantic==2.10.6
python-multipart==0.0.20
jinja2==3.1.6
python-dotenv==1.0.1
reportlab==4.4.2
pillow==10.4.0
openpyxl==3.1.2
psycopg2-binary==2.9.9